fix(workflow): update apparmor configuration process
Refactor the apparmor configuration setup in the vorpal workflow. Instead of using 'sudo cat' to write directly to '/etc/apparmor.d/bwrap', the content is first written to a temporary file 'bwrap' and then moved to the target directory with 'sudo mv'. This change ensures proper file permissions and avoids potential issues with direct file writing.