Implement simple registry (#97)
* feat: adds system to hash generation to avoid cross-platform conflicts

* feat: added basic registry service for artifacts

* refactor: simplify message handling and allow clippy lint

- Refactored message handling by replacing `if let None = message` with `if message.is_none()` for clarity and conciseness.
- Added `#[allow(clippy::too_many_arguments)]` attribute to suppress clippy lint warning for functions with many arguments.
- Removed unnecessary `return` statement after logging an error in `ArtifactService` implementation.

* refactor(artifact): merge mod.rs into artifact.rs and remove build.rs

- Renamed `mod.rs` to `artifact.rs` for better file organization.
- Removed `build.rs` as it is no longer needed.
- Cleaned up unused module import in `artifact.rs`.

* feat: fix environment variables passed to command

* feat: update vorpal workflow to configure apparmor for bwrap

- Added configuration for apparmor to allow unconfined execution of bwrap on Ubuntu runners.
- Created a new apparmor profile for bwrap and included site-specific additions.
- Restarted apparmor service to apply the new profile changes.

* fix: add sudo to apparmor configuration command in workflow script

The apparmor configuration command in the workflow script now uses sudo to ensure proper permissions when writing to /etc/apparmor.d/bwrap.

* fix(workflow): update apparmor configuration process

Refactor the apparmor configuration setup in the vorpal workflow. Instead of using 'sudo cat' to write directly to '/etc/apparmor.d/bwrap', the content is first written to a temporary file 'bwrap' and then moved to the target directory with 'sudo mv'. This change ensures proper file permissions and avoids potential issues with direct file writing.

* feat: added s3 backend support for registry

* refactor(cli, registry): simplify S3 bucket validation and default handling

- Combined the S3 backend check and bucket parameter validation into a single conditional statement in `main.rs`.
- Replaced `unwrap_or_else` with `unwrap_or_default` for cleaner default handling of the S3 bucket name in `service.rs`.

* feat(workflow): add AWS environment variables for S3 registry backend

Added AWS_ACCESS_KEY_ID, AWS_DEFAULT_REGION, and AWS_SECRET_ACCESS_KEY as environment variables to the workflow step that starts Vorpal with the S3 registry backend. This change ensures that the necessary AWS credentials are available for accessing the S3 bucket.

* feat(README)

* fix(paths): prevent setting timestamps on symlinks

This change adds a check to ensure that file timestamps are not set on symbolic links. This prevents potential issues with symlinks by only setting timestamps on regular files.

* fix: properly set dates on paths that are symlinks