Configuration language toolchains support (#105)
* feat: added rust language support in cli

* feat: add support for config toolchains

* feat: added language support for rust configurations

* refactor: remove protoc installation comment and update artifact names

- Removed the TODO comment for installing protoc in the GitHub Actions workflow.
- Updated the artifact name from 'vorpal-shell' to 'vorpal-dev' in the main.rs file.
- Changed the artifact_bin_paths from a vector to an array in the rust.rs file.

* feat: properly sets permissions and unpacks sources at build time

* feat: code improvements for building toolchains

* refactor(cli): simplify get_order function signature

Remove unnecessary lifetime annotation in the get_order function signature for improved code readability and maintainability.

* feat: add support for AWS SSO and Rustls

- Added dependencies for `aws-sdk-sso` and `aws-sdk-ssooidc` in `Cargo.lock`.
- Updated `aws-config` in `Cargo.toml` to include features `rt-tokio`, `rustls`, and `sso`.
- Introduced new packages such as `ring`, `rustls`, `rustls-native-certs`, `rustls-pemfile`, `rustls-webpki`, `hyper-rustls`, and others to support Rustls.
- Updated existing dependencies and checksums to accommodate new features.
- Enhanced security and networking capabilities with additional packages like `openssl-probe`, `schannel`, and `security-framework`.

* feat: move sanitize for artifacts after pack

* fix: move sanitize back

* fix(rust): correct directory name for update hashes

The directory name for storing update hashes was incorrect. Changed from "update-hash" to "update-hashes" to ensure consistency and correctness in the file structure.

* fix(cli): remove unnecessary newline in main.rs

feat(sdk): add symbolic link to /tmp in rust toolchain configuration

fix(worker): ensure return after error logging in artifact service implementation

* refactor(rust): remove unnecessary directory creation and symlink

Removed the creation of the 'update-hashes' directory and the symlink to '/tmp' in the rust toolchain configuration. Added 'auto_self_update' setting to 'settings.toml'.

* feat: support downloading remote sources

* fix: lint

* fix(protoc): update hash for X8664Linux target

Updated the hash value for the X8664Linux target in the protoc toolchain configuration to ensure correct artifact identification.

* fix(toolchain): update hash for X8664Linux in cargo.rs

Updated the hash value for the X8664Linux target in the cargo.rs file to ensure correct artifact retrieval.

* fix(toolchain): update hash for X8664Linux in clippy.rs

Updated the hash value for the X8664Linux target in the clippy.rs configuration file to ensure correct artifact retrieval.

* fix(toolchain): update hash for X8664Linux in rust_analyzer.rs

Updated the hash value for the X8664Linux target in the rust_analyzer configuration to ensure correct artifact retrieval.

* feat: implemented ArtifactSourceId support

* feat: implement exists endpoint

* refactor(rust_analyzer): improve import organization

Reorganize imports in rust_analyzer.rs for better readability and maintainability. No functional changes made.

* feat(paths): add cache directory setup in setup_paths function

Added logic to create a cache directory if it does not exist in the setup_paths function. This ensures that the cache directory is available for use, preventing potential errors related to missing directories.

* feat: checks before pushing sources to the registry

* feat: update hashes for x86_64

* fix: caching check for sources

* feat: update README and replace architecture diagram

- Updated the README section title from "Design" to "Infrastructure".
- Replaced the architecture diagram image from PNG to SVG format.
- Deleted the old PNG diagram file and added the new SVG diagram file.

* fix(readme): update image link to correct file name in diagram section

* refactor(logging): update log messages for consistency

Updated log messages to use a consistent format with '->' for better readability and uniformity across the codebase. Removed redundant log messages to streamline the logging process.

* chore: clean up toml files